VoiceAssistant/requirements_linux.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
geotext==0.4.0
googletrans==2.4.0
idna==2.9
pyalsaaudio==0.8.4
PyAudio==0.2.11
pyttsx3==2.87
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
urllib3==1.25.9
validators==0.14.3